refactor: convert file to ESM
- convert from cjs to esm
- replace ava with vitest
- run tests concurrently
